+++ app/routes/sketch+/artboards+/$slug_+/services/artboard/design/move-up.service.ts
@@ -7,7 +7,7 @@ import {
 } from '#app/models/design.server'
 import { type designTypeEnum } from '#app/schema/design'
 import { prisma, type IArtboard } from '#app/utils/db.server'
-import { artboardUpdateSelectedDesignService } from '../update-selected-design.service'
+import { artboardUpdateSelectedDesignService } from './update-selected.service'
 
 export const artboardDesignMoveUpService = async ({
 	userId,
