+++ app/routes/sketch+/artboards+/$slug_+/services/artboard/design/update-selected.service.ts
