+++ app/routes/sketch+/artboards+/$slug_+/services/layer/design/move-up.service.ts
@@ -8,7 +8,7 @@ import {
 import { type ILayer } from '#app/models/layer.server'
 import { type designTypeEnum } from '#app/schema/design'
 import { prisma } from '#app/utils/db.server'
-import { layerUpdateSelectedDesignService } from '../update-selected-design.service'
+import { layerUpdateSelectedDesignService } from './update-selected.service'
 
 export const layerDesignMoveUpService = async ({
 	userId,
