+++ app/routes/sketch+/artboards+/$slug_+/services/layer/design/toggle-visible.service.ts
@@ -7,7 +7,7 @@ import {
 import { type ILayer } from '#app/models/layer.server'
 import { type designTypeEnum } from '#app/schema/design'
 import { prisma } from '#app/utils/db.server'
-import { layerUpdateSelectedDesignService } from '../update-selected-design.service'
+import { layerUpdateSelectedDesignService } from './update-selected.service'
 
 export const layerDesignToggleVisibleService = async ({
 	userId,
